requirements.txt
Django==1.9.1
django-appconf==1.0.1
Pillow==3.1.0
Pygments==2.1
six==1.10.0
wheel==0.24.0
